feat(philo): finish implementing general skeleton

+++ milestone-3-philo/src/init_ctx.c
@@ -6,7 +6,7 @@
 /*   By: elagouch <elagouch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/09 16:07:16 by elagouch          #+#    #+#             */
-/*   Updated: 2025/05/09 17:15:40 by elagouch         ###   ########.fr       */
+/*   Updated: 2025/05/09 17:32:33 by elagouch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,8 +24,9 @@ t_ctx	*init_ctx(int argc, char **argv)
 	ctx->sleep_time = (unsigned long)ft_atoi(argv[4]);
 	if (argc == 6)
 		ctx->meals_count = (unsigned long)ft_atoi(argv[5]);
-	if (ctx->philos_count <= 0 || ctx->death_time <= 0 || ctx->meal_time <= 0
-		|| ctx->sleep_time <= 0 || ctx->meals_count < 0)
+	if ((int)ctx->philos_count <= 0 || (int)ctx->death_time <= 0
+		|| (int)ctx->meal_time <= 0 || (int)ctx->sleep_time <= 0
+		|| (int)ctx->meals_count < 0)
 	{
 		ft_printf_fd(STDERR_FILENO,
 			"Error: All arguments except the meals count must be"
